requirements.txt
Django==3.2.3
django-allauth==0.44.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.12.0
django-notifications-hq==1.6.0
django-plotly-dash==1.6.4
django-rest-multiple-models==2.1.3
django-taggit==1.4.0
django-widget-tweaks==1.4.1

